pazienteService, pazienteDTO

+++ src/main/java/it/capstone/arno/repository/RicoveroRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RicoveroRepository extends CrudRepository<Ricovero, Long> {
+public interface RicoveroRepository extends CrudRepository<Ricovero, Integer> {
 }
