+++ src/main/java/it/capstone/arno/repository/TerapiaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TerapiaRepository extends JpaRepository<Terapia, Long> {
+public interface TerapiaRepository extends JpaRepository<Terapia, Integer> {
 }
